Add more parameters checks
Perform a type and NULL check on the passed parameters for the parts
of the public API that were still lacking those.

Also, non-NULL checks on GObjects should really include a type check
as well: GTypes are runtime types and we cannot rely on g++ to catch
bad stuff for us.