box2d: Make simulation steps a constant length
Scaling the number of iterations up with the time doesn't work, as there
isn't a linear relation between iterations and step length. Instead, use
a constant step length (60fps) and run this multiple times with the same
amount of iterations to keep up with the timeline.

There is also a max step-time (15fps). If the frame-rate slows down beyond
this, the simulation will slow down too. This is to avoid possible
infinite loops when the CPU can't keep up, and to keep the frame-rate at
least 15fps.

This change keeps the simulation a lot more stable and allows us to
reduce the amount of iterations necessary to keep the simulation stable.